Created OpenResponseAuthoring component. Moved authoring component preview from the component authoring html to the node.html. #2854
