+++ src/main/webapp/site/src/app/teacher-hybrid-angular.module.ts
@@ -39,6 +39,7 @@ import { RubricAuthoringComponent } from '../../../wise5/authoringTool/rubric/ru
 import { NavItemProgressComponent } from './classroom-monitor/nav-item-progress/nav-item-progress.component';
 import { EditHTMLAdvancedComponent } from '../../../wise5/components/html/edit-html-advanced/edit-html-advanced.component';
 import { EditOutsideUrlAdvancedComponent } from '../../../wise5/components/outsideURL/edit-outside-url-advanced/edit-outside-url-advanced.component';
+import { OpenResponseAuthoring } from '../../../wise5/components/openResponse/open-response-authoring/open-response-authoring.component';
 
 @NgModule({
   declarations: [
@@ -61,6 +62,7 @@ import { EditOutsideUrlAdvancedComponent } from '../../../wise5/components/outsi
     NavItemProgressComponent,
     NodeAdvancedGeneralAuthoringComponent,
     NodeAdvancedJsonAuthoringComponent,
+    OpenResponseAuthoring,
     RubricAuthoringComponent,
     StatusIconComponent,
     WorkgroupInfoComponent,
