+++ src/main/webapp/wise5/authoringTool/node/node.html
@@ -188,6 +188,12 @@
       <summary-authoring ng-switch-when="Summary" node-id='{{nodeAuthoringController.nodeId}}' component-id='{{component.id}}'></summary-authoring>
       <table-authoring ng-switch-when="Table" node-id='{{nodeAuthoringController.nodeId}}' component-id='{{component.id}}'></table-authoring>
     </ng-content>
+    <div ng-style='{"border": "5px solid black"}'>
+      <div ng-style='{"padding-top": "20px", "padding-left": "20px"}'>
+        <h5>{{ ::'studentPreview' | translate }}</h5>
+      </div>
+      <preview-component node-id='{{nodeAuthoringController.nodeId}}' component-content='component'/>
+    </div>
     <md-divider ng-if='!$last'></md-divider>
   </div>
 </div>
