+++ src/main/webapp/wise5/components/animation/authoring.html
@@ -509,10 +509,4 @@
       </div>
     </div>
   </div>
-  <div ng-style='{"border": "5px solid black", "padding": "20px"}'>
-    <div>
-      <h5>{{ ::'studentPreview' | translate }}</h5>
-    </div>
-    <preview-component component-content='animationController.componentContent'/>
-  </div>
 </div>
