+++ src/main/webapp/wise5/components/audioOscillator/authoring.html
@@ -190,10 +190,4 @@
       </md-input-container>
     </div>
   </div>
-  <div ng-style='{"border": "5px solid black", "padding": "20px"}'>
-    <div>
-      <h5>{{ 'studentPreview' | translate }}</h5>
-    </div>
-    <preview-component component-content='audioOscillatorController.componentContent'/>
-  </div>
 </div>
