+++ src/main/webapp/wise5/components/conceptMap/authoring.html
@@ -429,11 +429,4 @@
       </md-button>
     </div>
   </div>
-  <div ng-style='{"border": "5px solid black"}'>
-    <div ng-style='{"padding-top": "20px", "padding-left": "20px"}'>
-      <h5>{{ ::'studentPreview' | translate }}</h5>
-    </div>
-    <preview-component component-content='conceptMapController.componentContent'
-        node-id='{{conceptMapController.nodeId}}'/>
-  </div>
 </div>
