+++ src/main/webapp/wise5/components/discussion/authoring.html
@@ -25,10 +25,4 @@
       </md-checkbox>
     </md-input-container>
   </div>
-  <div ng-style='{"border": "5px solid black", "padding": "20px"}'>
-    <div>
-      <h5>{{ ::'studentPreview' | translate }}</h5>
-    </div>
-    <preview-component component-content='discussionController.componentContent'/>
-  </div>
 </div>
