+++ src/main/webapp/wise5/components/discussion/discussionAuthoring.ts
@@ -36,23 +36,6 @@ class DiscussionAuthoringController extends EditComponentController {
     );
   }
 
-  connectedComponentTypeChanged(connectedComponent) {
-    this.changeAllDiscussionConnectedComponentTypesToMatch(connectedComponent.type);
-    super.connectedComponentTypeChanged(connectedComponent);
-  }
-
-  changeAllDiscussionConnectedComponentTypesToMatch(connectedComponentType) {
-    for (const connectedComponent of this.authoringComponentContent.connectedComponents) {
-      connectedComponent.type = connectedComponentType;
-    }
-  }
-
-  automaticallySetConnectedComponentTypeIfPossible(connectedComponent) {
-    if (connectedComponent.componentId != null) {
-      const firstConnectedComponent = this.authoringComponentContent.connectedComponents[0];
-      connectedComponent.type = firstConnectedComponent.type;
-    }
-  }
 }
 
 const DiscussionAuthoring = {
