+++ src/main/webapp/wise5/components/discussion/edit-discussion-advanced/edit-discussion-advanced.component.ts
@@ -2,6 +2,24 @@ import { EditAdvancedComponentAngularJSController } from "../../../../site/src/a
 
 class EditDiscussionAdvancedController extends EditAdvancedComponentAngularJSController {
   allowedConnectedComponentTypes = ['Discussion'];
+  
+  connectedComponentTypeChanged(connectedComponent) {
+    this.changeAllDiscussionConnectedComponentTypesToMatch(connectedComponent.type);
+    super.connectedComponentTypeChanged(connectedComponent);
+  }
+
+  changeAllDiscussionConnectedComponentTypesToMatch(connectedComponentType) {
+    for (const connectedComponent of this.authoringComponentContent.connectedComponents) {
+      connectedComponent.type = connectedComponentType;
+    }
+  }
+
+  automaticallySetConnectedComponentTypeIfPossible(connectedComponent) {
+    if (connectedComponent.componentId != null) {
+      const firstConnectedComponent = this.authoringComponentContent.connectedComponents[0];
+      connectedComponent.type = firstConnectedComponent.type;
+    }
+  }
 }
 
 export const EditDiscussionAdvancedComponent = {
