+++ src/main/webapp/wise5/components/draw/authoring.html
@@ -363,11 +363,4 @@
       </md-button>
     </div>
   </div>
-  <div ng-style='{"border": "5px solid black"}'>
-    <div ng-style='{"padding-top": "20px", "padding-left": "20px"}'>
-      <h5>{{ ::'studentPreview' | translate }}</h5>
-    </div>
-    <preview-component component-content='drawController.componentContent'
-        node-id='{{drawController.nodeId}}'/>
-  </div>
 </div>
