+++ src/main/webapp/wise5/components/draw/drawAuthoring.ts
@@ -251,47 +251,6 @@ class DrawAuthoringController extends EditComponentController {
     this.authoringComponentContent.stamps.Stamps[index] = fileName;
   }
 
-  automaticallySetConnectedComponentComponentIdIfPossible(connectedComponent) {
-    let numberOfAllowedComponents = 0;
-    let allowedComponent = null;
-    for (const component of this.getComponentsByNodeId(connectedComponent.nodeId)) {
-      if (
-        this.isConnectedComponentTypeAllowed(component.type) &&
-        component.id != this.componentId
-      ) {
-        numberOfAllowedComponents += 1;
-        allowedComponent = component;
-      }
-    }
-    if (numberOfAllowedComponents === 1) {
-      connectedComponent.componentId = allowedComponent.id;
-      connectedComponent.type = 'importWork';
-      this.setImportWorkAsBackgroundIfApplicable(connectedComponent);
-    }
-  }
-
-  connectedComponentComponentIdChanged(connectedComponent) {
-    connectedComponent.type = 'importWork';
-    this.setImportWorkAsBackgroundIfApplicable(connectedComponent);
-    this.authoringViewComponentChanged();
-  }
-
-  setImportWorkAsBackgroundIfApplicable(connectedComponent) {
-    const componentType = this.ProjectService.getComponentType(connectedComponent.nodeId,
-        connectedComponent.componentId);
-    if (['ConceptMap', 'Embedded', 'Graph', 'Label', 'Table'].includes(componentType)) {
-      connectedComponent.importWorkAsBackground = true;
-    } else {
-      delete connectedComponent.importWorkAsBackground;
-    }
-  }
-
-  importWorkAsBackgroundClicked(connectedComponent) {
-    if (!connectedComponent.importWorkAsBackground) {
-      delete connectedComponent.importWorkAsBackground;
-    }
-    this.authoringViewComponentChanged();
-  }
 }
 
 const DrawAuthoring = {
