+++ src/main/webapp/wise5/components/embedded/authoring.html
@@ -155,9 +155,5 @@
         </div>
       </div>
     </div>
-    <div ng-style='{"border": "5px solid black", "padding": "20px"}'>
-      <div><h5>{{ ::'studentPreview' | translate }}</h5></div>
-      <preview-component component-content='embeddedController.componentContent'/>
-    </div>
   </div>
 </div>
