+++ src/main/webapp/wise5/components/graph/authoring.html
@@ -805,11 +805,5 @@
         </div>
       </div>
     </div>
-    <div ng-style='{"border": "5px solid black"}'>
-      <div ng-style='{"padding-top": "20px", "padding-left": "20px"}'>
-        <h5>{{ ::'studentPreview' | translate }}</h5>
-      </div>
-      <preview-component component-content='graphController.componentContent'/>
-    </div>
   </div>
 </div>
