+++ src/main/webapp/wise5/components/graph/edit-graph-advanced/edit-graph-advanced.component.html
@@ -327,8 +327,7 @@
         <md-input-container style='margin-right: 20px;'>
           <md-checkbox class='md-primary'
                        ng-model='connectedComponent.importWorkAsBackground'
-                       ng-change='$ctrl.importWorkAsBackgroundClicked(connectedComponent)'
-                       ng-disabled='true'>
+                       ng-change='$ctrl.importWorkAsBackgroundClicked(connectedComponent)'>
             {{ ::'importWorkAsBackground' | translate }}
           </md-checkbox>
         </md-input-container>
