+++ src/main/webapp/wise5/components/graph/graphAuthoring.ts
@@ -365,114 +365,6 @@ class GraphAuthoringController extends EditComponentController {
     series.data = convertedData;
   }
 
-  addConnectedComponent() {
-    this.addConnectedComponentAndSetComponentIdIfPossible();
-    if (this.authoringComponentContent.connectedComponents.length > 1 ||
-        this.authoringComponentContent.series.length > 0) {
-      /*
-       * there is more than one connected component so we will enable
-       * trials so that each connected component can put work in a
-       * different trial
-       */
-      this.authoringComponentContent.enableTrials = true;
-    }
-    this.authoringViewComponentChanged();
-  }
-
-  addConnectedComponentSeriesNumber(connectedComponent) {
-    if (connectedComponent.seriesNumbers == null) {
-      connectedComponent.seriesNumbers = [];
-    }
-    connectedComponent.seriesNumbers.push(null);
-    this.authoringViewComponentChanged();
-  }
-
-  deleteConnectedComponentSeriesNumber(connectedComponent, seriesNumberIndex) {
-    if (connectedComponent.seriesNumbers == null) {
-      connectedComponent.seriesNumbers = [];
-    }
-    connectedComponent.seriesNumbers.splice(seriesNumberIndex, 1);
-    this.authoringViewComponentChanged();
-  }
-
-  connectedComponentSeriesNumberChanged(connectedComponent, seriesNumberIndex, value) {
-    if (connectedComponent.seriesNumbers == null) {
-      connectedComponent.seriesNumbers = [];
-    }
-    if (seriesNumberIndex < connectedComponent.seriesNumbers.length) {
-      connectedComponent.seriesNumbers[seriesNumberIndex] = value;
-    }
-    this.authoringViewComponentChanged();
-  }
-
-  connectedComponentComponentIdChanged(connectedComponent) {
-    const connectedComponentType = this.getConnectedComponentType(connectedComponent);
-    if (connectedComponentType !== 'Embedded') {
-      delete connectedComponent.seriesNumbers;
-    }
-    if (connectedComponentType !== 'Table') {
-      delete connectedComponent.skipFirstRow;
-      delete connectedComponent.xColumn;
-      delete connectedComponent.yColumn;
-    }
-    if (connectedComponentType !== 'Graph') {
-      delete connectedComponent.showClassmateWorkSource;
-    }
-    if (connectedComponentType === 'Table') {
-      connectedComponent.skipFirstRow = true;
-      connectedComponent.xColumn = 0;
-      connectedComponent.yColumn = 1;
-    }
-    connectedComponent.type = 'importWork';
-    this.setImportWorkAsBackgroundIfApplicable(connectedComponent);
-    this.authoringViewComponentChanged();
-  }
-
-  connectedComponentShowClassmateWorkChanged(connectedComponent) {
-    if (connectedComponent.showClassmateWork) {
-      connectedComponent.showClassmateWorkSource = 'period';
-    } else {
-      delete connectedComponent.showClassmateWork;
-      delete connectedComponent.showClassmateWorkSource;
-    }
-    this.authoringViewComponentChanged();
-  }
-
-  setImportWorkAsBackgroundIfApplicable(connectedComponent) {
-    const componentType = this.getConnectedComponentType(connectedComponent);
-    if (['ConceptMap','Draw','Label'].includes(componentType)) {
-      connectedComponent.importWorkAsBackground = true;
-    } else {
-      delete connectedComponent.importWorkAsBackground;
-    }
-  }
-
-  connectedComponentTypeChanged(connectedComponent) {
-    if (connectedComponent.type === 'importWork') {
-      delete connectedComponent.showClassmateWorkSource;
-    } else if (connectedComponent.type === 'showWork') {
-      delete connectedComponent.showClassmateWorkSource;
-    } else if (connectedComponent.type === 'showClassmateWork') {
-      /*
-       * the type has changed to show classmate work so we will enable
-       * trials so that each classmate work will show up in a
-       * different trial
-       */
-      this.authoringComponentContent.enableTrials = true;
-      if (connectedComponent.showClassmateWorkSource == null) {
-        connectedComponent.showClassmateWorkSource = 'period';
-      }
-    }
-    this.authoringViewComponentChanged();
-  }
-
-  importWorkAsBackgroundClicked(connectedComponent) {
-    if (!connectedComponent.importWorkAsBackground) {
-      delete connectedComponent.importWorkAsBackground;
-    }
-    this.authoringViewComponentChanged();
-  }
-
   enableMultipleYAxesChanged() {
     if (this.enableMultipleYAxes) {
       this.convertSingleYAxisToMultipleYAxes();
