+++ src/main/webapp/wise5/components/label/authoring.html
@@ -370,12 +370,4 @@
       </md-button>
     </div>
   </div>
-  <div ng-style='{"border": "5px solid black"}'>
-    <div ng-style='{"padding-top": "20px", "padding-left": "20px"}'>
-      <h5>{{ ::'studentPreview' | translate }}</h5>
-    </div>
-    <preview-component component-content='labelController.componentContent'
-        node-id='{{labelController.nodeId}}'/>
-  </div>
-</div>
 </div>
