+++ src/main/webapp/wise5/components/match/authoring.html
@@ -324,12 +324,4 @@
       </div>
     </div>
   </div>
-  <div ng-style='{"border": "5px solid black"}'>
-    <div ng-style='{"padding-top": "20px", "padding-left": "20px"}'>
-      <h5>{{ ::'studentPreview' | translate }}</h5>
-    </div>
-    <div>
-      <preview-component component-content='matchController.componentContent'/>
-    </div>
-  </div>
 </div>
