+++ src/main/webapp/wise5/components/multipleChoice/multipleChoiceAuthoring.ts
@@ -129,58 +129,6 @@ class MultipleChoiceAuthoringController extends EditComponentController {
     }
   }
 
-  automaticallySetConnectedComponentComponentIdIfPossible(connectedComponent) {
-    let numberOfAllowedComponents = 0;
-    let allowedComponent = null;
-    for (const component of this.getComponentsByNodeId(connectedComponent.nodeId)) {
-      if (this.isConnectedComponentTypeAllowed(component.type) &&
-        component.id != this.componentId
-      ) {
-        numberOfAllowedComponents += 1;
-        allowedComponent = component;
-      }
-    }
-    if (numberOfAllowedComponents === 1) {
-      connectedComponent.componentId = allowedComponent.id;
-      connectedComponent.type = 'importWork';
-      this.copyChoiceTypeAndChoicesFromConnectedComponent(connectedComponent);
-    }
-  }
-
-  connectedComponentComponentIdChanged(connectedComponent) {
-    connectedComponent.type = 'importWork';
-    this.copyChoiceTypeAndChoicesFromConnectedComponent(connectedComponent);
-    this.authoringViewComponentChanged();
-  }
-
-  copyChoiceTypeAndChoicesFromConnectedComponent(connectedComponent) {
-    const nodeId = connectedComponent.nodeId;
-    const componentId = connectedComponent.componentId;
-    if (
-      this.ProjectService.getComponentByNodeIdAndComponentId(nodeId, componentId).type ===
-      'MultipleChoice'
-    ) {
-      this.copyChoiceTypeFromComponent(nodeId, componentId);
-      this.copyChoicesFromComponent(nodeId, componentId);
-    }
-  }
-
-  copyChoiceTypeFromComponent(nodeId, componentId) {
-    const component = this.ProjectService.getComponentByNodeIdAndComponentId(nodeId, componentId);
-    this.authoringComponentContent.choiceType = component.choiceType;
-  }
-
-  copyChoicesFromComponent(nodeId, componentId) {
-    this.authoringComponentContent.choices = this.getCopyOfChoicesFromComponent(
-      nodeId,
-      componentId
-    );
-  }
-
-  getCopyOfChoicesFromComponent(nodeId, componentId) {
-    const component = this.ProjectService.getComponentByNodeIdAndComponentId(nodeId, componentId);
-    return this.UtilService.makeCopyOfJSONObject(component.choices);
-  }
 }
 
 const MultipleChoiceAuthoring = {
