--- src/main/webapp/wise5/components/openResponse/authoring.html
@@ -1,24 +1,0 @@
-<div class='openResponse' flex>
-  <style>
-    {{openResponseController.nodeContent.style}}
-  </style>
-  <div>
-    <div ng-if='!openResponseController.authoringComponentContent.showPreviousWork'>
-      <md-input-container style='width:100%'>
-        <label>{{ ::'PROMPT' | translate }}</label><span><md-icon>announcement</md-icon><md-tooltip md-direction='right'>{{ ::'openResponse.promptTip' | translate }}</md-tooltip></span>
-        <textarea rows='1'
-                  ng-model='openResponseController.authoringComponentContent.prompt'
-                  ng-change='openResponseController.authoringViewComponentChanged()'
-                  ng-model-options='{ debounce: 1000 }'
-                  placeholder='{{ "enterPromptHere" | translate }}'>
-      </textarea>
-      </md-input-container>
-    </div>
-  </div>
-  <div ng-style='{"border": "5px solid black", "padding": "20px"}'>
-    <div>
-      <h5>{{ ::'studentPreview' | translate }}</h5>
-    </div>
-    <preview-component component-content='openResponseController.componentContent'/>
-  </div>
-</div>
