+++ src/main/webapp/wise5/components/openResponse/openResponseAuthoringComponentModule.ts
@@ -1,15 +1,16 @@
 'use strict';
 
 import * as angular from 'angular';
-import { downgradeInjectable } from '@angular/upgrade/static';
+import { downgradeComponent, downgradeInjectable } from '@angular/upgrade/static';
 import { OpenResponseService } from './openResponseService';
-import OpenResponseAuthoring from './openResponseAuthoring';
+import { OpenResponseAuthoring } from './open-response-authoring/open-response-authoring.component';
 import { EditOpenResponseAdvancedComponent } from './edit-open-response-advanced/edit-open-response-advanced.component';
 
 const openResponseAuthoringComponentModule = angular
   .module('openResponseAuthoringComponentModule', ['pascalprecht.translate'])
   .service('OpenResponseService', downgradeInjectable(OpenResponseService))
-  .component('openResponseAuthoring', OpenResponseAuthoring)
+  .directive('openResponseAuthoring', downgradeComponent(
+      { component: OpenResponseAuthoring }) as angular.IDirectiveFactory)
   .component('editOpenResponseAdvanced', EditOpenResponseAdvancedComponent)
   .config([
     '$translatePartialLoaderProvider',
