+++ src/main/webapp/wise5/components/outsideURL/authoring.html
@@ -104,4 +104,3 @@
     <md-icon>refresh</md-icon>
   </md-button>
 </div>
-<preview-component component-content='outsideURLController.componentContent'/>
