+++ src/main/webapp/wise5/components/summary/authoring.html
@@ -225,14 +225,4 @@
       </md-tooltip>
     </md-button>
   </div>
-  <div ng-style='{"border": "5px solid black"}'>
-    <div ng-style='{"padding-top": "20px", "padding-left": "20px"}'>
-      <h5>{{ ::'studentPreview' | translate }}</h5>
-    </div>
-    <div>
-      <compile id='prompt' data='summaryController.getPrompt()'></compile>
-      <div>
-        <preview-component component-content='summaryController.componentContent'/>
-      </div>
-  </div>
 </div>
