+++ src/main/webapp/wise5/components/table/authoring.html
@@ -378,11 +378,4 @@
         </div>
       </div>
     </div>
-    <br/>
-    <div ng-style='{"border": "5px solid black"}'>
-      <div ng-style='{"padding-top": "20px", "padding-left": "20px"}'>
-        <h5>{{ ::'studentPreview' | translate }}</h5>
-      </div>
-      <preview-component component-content='tableController.componentContent'/>
-    </div>
   </div>
